requirements.txt
block.bootstrap.pytorch
h5py
plotly==3.10.0
pyyaml==5.4.1